# go-server-client/go/model_example_parent_model.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Example API
 *
 * Example API spec
 *
 * API version: v1
 * Contact: example@example-team.com
 */

package openapi




// ExampleParentModel - Object representing example parent model. 
type ExampleParentModel struct {

	// Parent ID.
	ParentId string `json:"ParentId,omitempty"`

	// Parent name field.
	ParentName string `json:"ParentName,omitempty"`
}

// AssertExampleParentModelRequired checks if the required fields are not zero-ed
func AssertExampleParentModelRequired(obj ExampleParentModel) error {
	return nil
}

// AssertExampleParentModelConstraints checks if the values respects the defined constraints
func AssertExampleParentModelConstraints(obj ExampleParentModel) error {
	return nil
}